requirements.txt
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
bitsandbytes==0.48.1
blessed==1.22.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
decorator==5.2.1
decord==0.6.0
diffusers==0.35.2
dill==0.4.0
distlib==0.4.0
dreamsim==0.2.1
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
facexlib==0.3.0
filelock==3.19.1
filterpy==1.4.5
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gpustat==1.1.1
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
icecream==2.1.8
identify==2.6.15
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
iniconfig==2.3.0
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==1.7.2
mmengine==0.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.3
nodeenv==1.9.1
numba==0.62.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==3.2.0
openai-clip==1.0.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
Pygments==2.19.2
pyiqa==0.1.14.1
pyparsing==3.2.5
pyrallis==0.3.1
pytest==9.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
ruff==0.14.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.5.10
timm==1.0.20
tokenizers==0.22.1
tomli==2.3.0
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
